+++ CHANGELOG.md
@@ -8,12 +8,35 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased] - yyyy-mm-dd
 
 ### Added
+- 
+
+### Changed
+- 
+
+### Fixed
+-
+
+## [0.0.3] - 2025-01-09
+
+### Added
 - Added random enum selection through `Utilities.EnumExtensions.GetRandomEnum()`
+- Added `JMath.cs` to provide various math & number functions
+  - Includes option to convert `int` into a roman numeral
+- Added `LerpFunctions.cs` as utility for custom lerp behaviours
+- Added `PhysicsLauncher3D.cs` as a 3D option for launch objects
+  - Includes appropriate Gizmos as well
+- Added `MouseCaster.cs` into Utilities
 
 ### Changed
 - Updated README to include Samples tutorial
 - Updated `package.json` to show author information
-
+- Added `Draw.Circle()` with overload that includes the circles normal
+- Added `Shuffle()` extension method for Lists & Arrays in `CollectionExtensions.cs`
+- Set `CollectionExtensions.cs` to use `System.Random` instead of `UnityEngine.Random`
+- Added additional functions to `CollisionChecks.cs`
+  - `Line2Rect()` - Checks if a line is interacting a rectangle
+  - `Poly2Point()` - Checks if a point is interacting with a polygon 
+  - `Poly2Rect()` - Checks if a rectangle is interacting with a polygon
 ### Fixed
 -